統合テスト_エネミーの挙動テスト更新 ・プレイヤーが攻撃を受けた時にHPの減少を確認するテストを追加 ・上記テストを行なうため、各スクリプトをリファクター

+++ Assets/Script/AIFSM/EnemyCtrl.cs
@@ -183,7 +183,7 @@ namespace Cradle.FM{
 				}
 			}
 
-			public int GetHP(){
+			public float GetHP(){
 				return status.GetHP ();
 			}
